Change DNS used by PgBouncer to fix DNS lookup issues
We were previously using evdns2, and apparently that has issues
in docker. Now switched to libc-2.27 for DNS. Will see if this
fixes the core issue.

The symptom is that PgBouncer frequently failed to perform hostname
lookups on the "db" host. It has a 15 second timeout, thankfully,
so this issue ended up rearing it's head quickly and frequently.

See:
https://github.com/pgbouncer/pgbouncer/issues/122
https://github.com/pgbouncer/pgbouncer/issues/354